+++ src/CodeGeneration/Fixtures/definitionGroupWithCommandTypePropertiesFixture.php
@@ -18,9 +18,9 @@ final class DoSomething implements SerializablePayload
         return $this->reason;
     }
 
-    public static function fromPayload(array $payload): self
+    public static function fromPayload(array $payload): static
     {
-        return new DoSomething(
+        return new static(
             (string) $payload['reason']
         );
     }