+++ src/CodeGeneration/Fixtures/definitionGroupWithDefaultsTypePropertiesFixture.php
@@ -18,9 +18,9 @@ final class EventWithDescription implements SerializablePayload
         return $this->description;
     }
 
-    public static function fromPayload(array $payload): self
+    public static function fromPayload(array $payload): static
     {
-        return new EventWithDescription(
+        return new static(
             (string) $payload['description']
         );
     }