+++ src/CodeGeneration/Fixtures/groupWithEventWithNoFieldsTypePropertiesFixture.php
@@ -8,9 +8,9 @@ use EventSauce\EventSourcing\Serialization\SerializablePayload;
 
 final class WithoutFields implements SerializablePayload
 {
-    public static function fromPayload(array $payload): self
+    public static function fromPayload(array $payload): static
     {
-        return new WithoutFields();
+        return new static();
     }
 
     public function toPayload(): array